requirements.txt
discord.py==2.0.1
requests==2.24.0
pymongo[tls,srv]==4.2.0